Fix documentation of key name for specifiying map parent element
The option for setting the CSS selector for the map parent element is listed as `element`. However, using this throws the error `Uncaught TypeError: Cannot set property 'innerHTML' of null` with references to lines 19 and 70. I'm not exactly sure why `element` doesn't work, but using `el` instead does work. I'm suggesting updating the documentation to reflect that `el` should be the key name used to specify the map parent element.